Add files via upload
In order to study the regional impact of Soviet economic policy, I have digitised the birth rates and the employment in different oblasts as a fraction of their rural populations from 1950 to 1978. The data is digistised by using the Narkhoz website and is partially taken from Koropeckyj and Schroeder's work: Economics of Soviet Regions. As the variables are only available at an oblast level (in together 15 of them), they will be used as oblast level controls and not as the main explanatory variables.